Add support for .NET Standard 2.0
- Fix handling of library on Linux. Unfortunately, .NET Core doesn't support DllMap like Mono does, so we need to compile the P/Invokes twice (once for Windows, once for Linux) and have an ugly wrapper class as a facade
- Convert gpgme-sharp.csproj to new project file format
- Convert gpgme-sharp.sln to Visual Studio 2017 format
