+++ gpgme-sharp.sln
@@ -1,7 +1,9 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "gpgme-sharp", "gpgme-sharp\gpgme-sharp.csproj", "{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}"
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.271
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "gpgme-sharp", "gpgme-sharp\gpgme-sharp.csproj", "{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{F803B37B-3E35-43F8-A772-3A6D0ECB610A}"
 EndProject
@@ -29,6 +31,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PgpSignVerify", "Examples\P
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SignPgpKey", "Examples\SignPgpKey\SignPgpKey.csproj", "{B2B6890B-54C8-4BA7-AE61-F06E582A6A3C}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GPGME.Native.Win32", "GPGME.Native.Win32\GPGME.Native.Win32.csproj", "{601F32CB-CF0D-486B-848F-8A07BA42658F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GPGME.Native.Unix", "GPGME.Native.Unix\GPGME.Native.Unix.csproj", "{3001DC3D-E42D-4DD1-8960-32A9580FA94D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GPGME.Native.Shared", "GPGME.Native.Shared\GPGME.Native.Shared.csproj", "{C18682B5-3246-4ED4-9527-3F51E89A7235}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -37,14 +45,14 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Debug|Any CPU.Build.0 = Debug|x86
-		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Debug|x86.ActiveCfg = Debug|x86
-		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Debug|x86.Build.0 = Debug|x86
-		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Release|Any CPU.ActiveCfg = Release|x86
-		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Release|Any CPU.Build.0 = Release|x86
-		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Release|x86.ActiveCfg = Release|x86
-		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Release|x86.Build.0 = Release|x86
+		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Debug|x86.Build.0 = Debug|Any CPU
+		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Release|x86.ActiveCfg = Release|Any CPU
+		{5523357F-F3CA-4E25-B6EC-30EE3F5D769D}.Release|x86.Build.0 = Release|Any CPU
 		{81E13A5A-2BF9-4EFB-9AC8-FB004703B648}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{81E13A5A-2BF9-4EFB-9AC8-FB004703B648}.Debug|Any CPU.Build.0 = Debug|x86
 		{81E13A5A-2BF9-4EFB-9AC8-FB004703B648}.Debug|x86.ActiveCfg = Debug|x86
@@ -125,6 +133,30 @@ Global
 		{B2B6890B-54C8-4BA7-AE61-F06E582A6A3C}.Release|Any CPU.Build.0 = Release|x86
 		{B2B6890B-54C8-4BA7-AE61-F06E582A6A3C}.Release|x86.ActiveCfg = Release|x86
 		{B2B6890B-54C8-4BA7-AE61-F06E582A6A3C}.Release|x86.Build.0 = Release|x86
+		{601F32CB-CF0D-486B-848F-8A07BA42658F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{601F32CB-CF0D-486B-848F-8A07BA42658F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{601F32CB-CF0D-486B-848F-8A07BA42658F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{601F32CB-CF0D-486B-848F-8A07BA42658F}.Debug|x86.Build.0 = Debug|Any CPU
+		{601F32CB-CF0D-486B-848F-8A07BA42658F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{601F32CB-CF0D-486B-848F-8A07BA42658F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{601F32CB-CF0D-486B-848F-8A07BA42658F}.Release|x86.ActiveCfg = Release|Any CPU
+		{601F32CB-CF0D-486B-848F-8A07BA42658F}.Release|x86.Build.0 = Release|Any CPU
+		{3001DC3D-E42D-4DD1-8960-32A9580FA94D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3001DC3D-E42D-4DD1-8960-32A9580FA94D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3001DC3D-E42D-4DD1-8960-32A9580FA94D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3001DC3D-E42D-4DD1-8960-32A9580FA94D}.Debug|x86.Build.0 = Debug|Any CPU
+		{3001DC3D-E42D-4DD1-8960-32A9580FA94D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3001DC3D-E42D-4DD1-8960-32A9580FA94D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3001DC3D-E42D-4DD1-8960-32A9580FA94D}.Release|x86.ActiveCfg = Release|Any CPU
+		{3001DC3D-E42D-4DD1-8960-32A9580FA94D}.Release|x86.Build.0 = Release|Any CPU
+		{C18682B5-3246-4ED4-9527-3F51E89A7235}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C18682B5-3246-4ED4-9527-3F51E89A7235}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C18682B5-3246-4ED4-9527-3F51E89A7235}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C18682B5-3246-4ED4-9527-3F51E89A7235}.Debug|x86.Build.0 = Debug|Any CPU
+		{C18682B5-3246-4ED4-9527-3F51E89A7235}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C18682B5-3246-4ED4-9527-3F51E89A7235}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C18682B5-3246-4ED4-9527-3F51E89A7235}.Release|x86.ActiveCfg = Release|Any CPU
+		{C18682B5-3246-4ED4-9527-3F51E89A7235}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -132,15 +164,18 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{81E13A5A-2BF9-4EFB-9AC8-FB004703B648} = {F803B37B-3E35-43F8-A772-3A6D0ECB610A}
 		{040ADE11-7491-4C7C-9A4B-E4B645F08BD5} = {F803B37B-3E35-43F8-A772-3A6D0ECB610A}
+		{478F350D-4A32-4B68-BC3E-EC913CD097CF} = {040ADE11-7491-4C7C-9A4B-E4B645F08BD5}
 		{34B6A704-4985-40D1-8D8D-09C4C97227EA} = {F803B37B-3E35-43F8-A772-3A6D0ECB610A}
+		{E5A0AC77-765C-4E26-ADA1-C19D0B0D2745} = {34B6A704-4985-40D1-8D8D-09C4C97227EA}
+		{4A57F975-5C75-4454-933D-5F3EA8958D90} = {34B6A704-4985-40D1-8D8D-09C4C97227EA}
+		{6BACB1D8-F407-4701-86D3-323C878B0C31} = {34B6A704-4985-40D1-8D8D-09C4C97227EA}
 		{EAD627D1-7102-48BA-AA28-1F0E5B28876E} = {F803B37B-3E35-43F8-A772-3A6D0ECB610A}
 		{B4DC5583-0F25-4451-A15E-171CE7583D3E} = {F803B37B-3E35-43F8-A772-3A6D0ECB610A}
 		{CA80E147-9549-483F-AC09-57A46891BBC2} = {F803B37B-3E35-43F8-A772-3A6D0ECB610A}
 		{F593613E-08BF-4B4C-A863-4FFD631E369B} = {F803B37B-3E35-43F8-A772-3A6D0ECB610A}
 		{B2B6890B-54C8-4BA7-AE61-F06E582A6A3C} = {F803B37B-3E35-43F8-A772-3A6D0ECB610A}
-		{478F350D-4A32-4B68-BC3E-EC913CD097CF} = {040ADE11-7491-4C7C-9A4B-E4B645F08BD5}
-		{E5A0AC77-765C-4E26-ADA1-C19D0B0D2745} = {34B6A704-4985-40D1-8D8D-09C4C97227EA}
-		{4A57F975-5C75-4454-933D-5F3EA8958D90} = {34B6A704-4985-40D1-8D8D-09C4C97227EA}
-		{6BACB1D8-F407-4701-86D3-323C878B0C31} = {34B6A704-4985-40D1-8D8D-09C4C97227EA}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {76405F3E-1C3A-4FBE-9275-780388EAA45B}
 	EndGlobalSection
 EndGlobal
