+++ gpgme-sharp/Key.cs
@@ -79,7 +79,7 @@ namespace Libgpgme
 
         private void CleanUp() {
             if (_key_ptr != IntPtr.Zero) {
-                libgpgme.gpgme_key_release(_key_ptr);
+                libgpgme.NativeMethods.gpgme_key_release(_key_ptr);
                 _key_ptr = IntPtr.Zero;
             }
         }
@@ -143,7 +143,7 @@ namespace Libgpgme
                 _instance_key_edit_callback = KeyEditCallback;
 
                 // start key editing
-                int err = libgpgme.gpgme_op_edit(
+                int err = libgpgme.NativeMethods.gpgme_op_edit(
                     ctx.CtxPtr,
                     KeyPtr,
                     _instance_key_edit_callback,
@@ -186,7 +186,7 @@ namespace Libgpgme
         }
 
         protected virtual int KeyEditCallback(IntPtr handle, KeyEditStatusCode status, string args, int fd) {
-            libgpgme.gpgme_io_write(fd, new[] {(byte) 'q', (byte) 'u', (byte) 'i', (byte) 't', (byte) '\n'}, (UIntPtr)5);
+            libgpgme.NativeMethods.gpgme_io_write(fd, new[] {(byte) 'q', (byte) 'u', (byte) 'i', (byte) 't', (byte) '\n'}, (UIntPtr)5);
             throw new NotImplementedException("The function KeyEditCallback is not implemented.");
         }
     }
