+++ gpgme-sharp/PgpKey.cs
@@ -93,7 +93,7 @@ namespace Libgpgme
                                             ref passphrase);
                                     if (passphrase != null) {
                                         byte[] p = Gpgme.ConvertCharArrayToUTF8(passphrase, 0);
-                                        libgpgme.gpgme_io_write(fd, p, (UIntPtr)p.Length);
+                                        libgpgme.NativeMethods.gpgme_io_write(fd, p, (UIntPtr)p.Length);
 
                                         int i;
                                         // try to clear passphrase in memory
@@ -108,7 +108,7 @@ namespace Libgpgme
                                     byte[] p = Gpgme.ConvertCharArrayToUTF8(
                                         _settings.passSettings.Passphrase,
                                         0);
-                                    libgpgme.gpgme_io_write(fd, p, (UIntPtr)p.Length);
+                                    libgpgme.NativeMethods.gpgme_io_write(fd, p, (UIntPtr)p.Length);
 
                                     int i;
                                     // try to clear passphrase in memory
@@ -117,7 +117,7 @@ namespace Libgpgme
                                     }
                                 } else {
                                     // No password or password callback function specified!
-                                    libgpgme.gpgme_io_write(fd, new[] { (byte)0 }, (UIntPtr)1);
+                                    libgpgme.NativeMethods.gpgme_io_write(fd, new[] { (byte)0 }, (UIntPtr)1);
                                 }
 
                                 output = new byte[0]; // confirm password (send \n)
@@ -173,8 +173,8 @@ namespace Libgpgme
 #if (VERBOSE_DEBUG)
                 DebugOutput(output);
 #endif
-                libgpgme.gpgme_io_write(fd, output, (UIntPtr)output.Length);
-                libgpgme.gpgme_io_write(fd, new[] {(byte) '\n'}, (UIntPtr)1);
+                libgpgme.NativeMethods.gpgme_io_write(fd, output, (UIntPtr)output.Length);
+                libgpgme.NativeMethods.gpgme_io_write(fd, new[] {(byte) '\n'}, (UIntPtr)1);
             }
 
             return 0;
@@ -1097,7 +1097,7 @@ namespace Libgpgme
 
                             if (passphrase != null) {
                                 byte[] p = Gpgme.ConvertCharArrayToUTF8(passphrase, 0);
-                                libgpgme.gpgme_io_write(fd, p, (UIntPtr) p.Length);
+                                libgpgme.NativeMethods.gpgme_io_write(fd, p, (UIntPtr) p.Length);
 
                                 // try to clear passphrase in memory
                                 int i;
