--- gpgme-sharp/Properties/AssemblyInfo.cs
@@ -1,15 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("gpgme-sharp")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("gpgme-sharp")]
-[assembly: AssemblyCopyright("Copyright © 2009-2013 Daniel Mueller")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: ComVisible(false)]
-[assembly: Guid("2110b514-e57b-4f97-9d0f-80307774165b")]
-[assembly: AssemblyVersion("1.1.0.0")]
-[assembly: AssemblyFileVersion("1.1.0.0")]
